OXY-41 : Loan Repayment Reminders
