+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/service/LoanSchedularServiceImpl.java
@@ -18,8 +18,6 @@
  */
 package org.apache.fineract.portfolio.loanaccount.service;
 
-import static java.util.stream.Collectors.toSet;
-
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.security.SecureRandom;
 import java.util.ArrayList;
@@ -48,8 +46,7 @@ import org.apache.fineract.infrastructure.jobs.annotation.CronTarget;
 import org.apache.fineract.infrastructure.jobs.exception.JobExecutionException;
 import org.apache.fineract.infrastructure.jobs.service.JobName;
 import org.apache.fineract.infrastructure.security.service.PlatformSecurityContext;
-import org.apache.fineract.notification.data.NotificationData;
-import org.apache.fineract.notification.eventandlistener.NotificationEventPublisher;
+import org.apache.fineract.notification.service.ActiveMqNotificationDomainServiceImpl;
 import org.apache.fineract.organisation.office.data.OfficeData;
 import org.apache.fineract.organisation.office.exception.OfficeNotFoundException;
 import org.apache.fineract.organisation.office.service.OfficeReadPlatformService;
@@ -63,7 +60,6 @@ import org.apache.fineract.portfolio.loanaccount.domain.LoanRepository;
 import org.apache.fineract.portfolio.loanaccount.loanschedule.data.LoanRepaymentReminderData;
 import org.apache.fineract.portfolio.loanaccount.loanschedule.data.OverdueLoanScheduleData;
 import org.apache.fineract.useradministration.domain.AppUser;
-import org.apache.fineract.useradministration.domain.AppUserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -94,9 +90,9 @@ public class LoanSchedularServiceImpl implements LoanSchedularService {
     private final LoanRepaymentReminderSettingsRepository loanRepaymentReminderSettingsRepository;
     private final LoanRepaymentReminderRepository loanRepaymentReminderRepository;
     private final PlatformSecurityContext context;
-    private final NotificationEventPublisher notificationEventPublisher;
-    private final AppUserRepository appUserRepository;
     private final FromJsonHelper fromApiJsonHelper;
+    @Autowired
+    private ActiveMqNotificationDomainServiceImpl activeMqNotificationDomainService;
     @Autowired
     private Environment env;
 
@@ -381,9 +377,9 @@ public class LoanSchedularServiceImpl implements LoanSchedularService {
                 if (!CollectionUtils.isEmpty(reminders)) {
                     for (LoanRepaymentReminder data : reminders) {
                         LoanMessageRepaymentReminderData repaymentReminderData = new LoanMessageRepaymentReminderData(data);
-                        buildNotification("ALL_FUNCTION", "LoanRepaymentReminder", data.getId(),
+                        activeMqNotificationDomainService.buildNotification("ALL_FUNCTION", "LoanRepaymentReminder", data.getId(),
                                 this.fromApiJsonHelper.toJson(repaymentReminderData), "PENDING", context.authenticatedUser().getId(),
-                                officeId);
+                                officeId,this.env.getProperty("fineract.activemq.loanRepaymentReminderQueue"));
                     }
 
                 } else {
@@ -401,26 +397,6 @@ public class LoanSchedularServiceImpl implements LoanSchedularService {
         return user;
     }
 
-    private void buildNotification(String permission, String objectType, Long objectIdentifier, String notificationContent,
-            String eventType, Long appUserId, Long officeId) {
 
-        String tenantIdentifier = ThreadLocalContextUtil.getTenant().getTenantIdentifier();
-        Set<Long> userIds = getNotifiableUserIds(officeId, permission);
-        NotificationData notificationData = new NotificationData(objectType, objectIdentifier, eventType, appUserId, notificationContent,
-                false, false, tenantIdentifier, officeId, userIds);
-        try {
-            notificationEventPublisher.broadcastGenericActiveMqNotification(notificationData,
-                    this.env.getProperty("fineract.activemq.loanRepaymentReminderQueue"));
-        } catch (Exception e) {
-            // We want to avoid rethrowing the exception to stop the business transaction from rolling back
-            log.error("Error while broadcasting notification event", e);
-        }
-    }
-
-    private Set<Long> getNotifiableUserIds(Long officeId, String permission) {
-        Collection<AppUser> users = appUserRepository.findByOfficeId(officeId);
-        Collection<AppUser> usersWithPermission = users.stream().filter(aU -> aU.hasAnyPermission(permission, "ALL_FUNCTIONS")).toList();
-        return usersWithPermission.stream().map(AppUser::getId).collect(toSet());
-    }
 
 }
