+++ fineract-provider/src/main/resources/application.properties
@@ -151,4 +151,5 @@ spring.main.allow-bean-definition-overriding=true
 
 fiter.features.ff4j=${FF4J_FILE_NAME:ff4j-feature-template.yml}
 fineract.activemq.loanRepaymentReminderQueue=${LOAN_REPAYMENT_REMINDER_QUEUE:LoanRepaymentReminders}
+fineract.activemq.loanRepaymentConfirmationQueue=${LOAN_REPAYMENT_CONFIRMATION_QUEUE:LoanRepaymentConfirmation}
 fineract.activemq.blockerUrl=${ACTIVEMQ_BLOCKER_URL:tcp://localhost:61616}
